fix: ReadingSpaceMembers 객체의 카운트 조회 로직 추가
- 기존에는 ReadingSpaceMemberRole의 데이터 전체를 다 불러와서 카운팅 조회를 하도록 구현을 함. 하지만 만약 데이터가 엄청 많은 경우, 전체를 조회해서 데이터를 가져온 다음에 size() 메소드로 카운팅을 진행하기 때문에 성능의 문제가 발생할 수 있음
- ReadingSpaceMemberRole에 대해 카운트 조회만 하는 기능을 추가해 ReadingSpace에 멤버가 2명 이상일 때는 삭제가 불가능하다는 예외를 던지게 하도록 변경함. 따라서 삭제 기능에서 조회되는 데이터가 많은 경우, 카운팅 조회만 하고 메소드가 종료되게 만들었음

+++ src/main/java/com/sh/threesentences/readingspace/service/ReadingSpaceService.java
@@ -1,12 +1,13 @@
 package com.sh.threesentences.readingspace.service;
 
+import static com.sh.threesentences.readingspace.exception.ReadingSpaceErrorCode.MEMBER_IS_STILL_IN_SPACE;
 import static com.sh.threesentences.readingspace.exception.ReadingSpaceErrorCode.READING_SPACE_NOT_FOUND;
 
 import com.sh.threesentences.common.enums.OpenType;
 import com.sh.threesentences.readingspace.dto.ReadingSpaceRequestDto;
 import com.sh.threesentences.readingspace.dto.ReadingSpaceResponseDto;
 import com.sh.threesentences.readingspace.entity.ReadingSpace;
-import com.sh.threesentences.readingspace.entity.UserReadingSpaceMapping;
+import com.sh.threesentences.readingspace.entity.ReadingSpaceMemberRole;
 import com.sh.threesentences.readingspace.entity.fcc.ReadingSpaceMembers;
 import com.sh.threesentences.readingspace.enums.UserRole;
 import com.sh.threesentences.readingspace.repository.ReadingSpaceRepository;
@@ -31,9 +32,9 @@ public class ReadingSpaceService {
         ReadingSpace savedReadingSpace = readingSpaceRepository.save(readingSpaceRequestDto.toEntity());
 
         // TODO: 로그인한 사용자의 정보를 조회해서 User 엔티티를 UserReadingSpaceMapping에 추가해줘야함.
-        UserReadingSpaceMapping userReadingSpaceMapping = UserReadingSpaceMapping.createUserReadingSpaceMapping(user,
+        ReadingSpaceMemberRole readingSpaceMemberRole = new ReadingSpaceMemberRole(user,
             savedReadingSpace, UserRole.ADMIN);
-        userReadingSpaceRepository.save(userReadingSpaceMapping);
+        userReadingSpaceRepository.save(readingSpaceMemberRole);
 
         return ReadingSpaceResponseDto.fromEntity(savedReadingSpace);
     }
@@ -59,18 +60,28 @@ public class ReadingSpaceService {
         Long id = 1L;
         return userReadingSpaceRepository.findByUserId(id)
             .stream()
-            .map(UserReadingSpaceMapping::getReadingSpace)
+            .map(ReadingSpaceMemberRole::getReadingSpace)
             .map(ReadingSpaceResponseDto::fromEntity)
             .collect(Collectors.toList());
     }
 
     public void delete(Long id) {
         // TODO: 사용자 정보 조회 후, 어드민 권한인 경우에만 삭제 기능을 진행할 수 있게 변경
-        List<UserReadingSpaceMapping> membersOfReadingSpace = userReadingSpaceRepository.findByReadingSpaceId(id);
+
+        int memberCount = userReadingSpaceRepository.countByReadingSpaceId(id);
+        if (hasMembersGreaterThanOne(memberCount)) {
+            throw new IllegalStateException(MEMBER_IS_STILL_IN_SPACE.getMessage());
+        }
+
+        List<ReadingSpaceMemberRole> membersOfReadingSpace = userReadingSpaceRepository.findByReadingSpaceId(id);
         ReadingSpaceMembers readingSpaceMembers = ReadingSpaceMembers.fromEntity(membersOfReadingSpace);
         readingSpaceMembers.checkSpaceDeleteCondition();
 
         ReadingSpace readingSpace = membersOfReadingSpace.get(0).getReadingSpace();
         readingSpace.delete();
+    }
+
+    private boolean hasMembersGreaterThanOne(int memberCount) {
+        return memberCount >= 2;
     }
 }
